Basic trick functionality complete.
Two clocks, the normal, and trick clock. The former simply increments once a
real second, acting as a normal clock would. The latter independently
increments twice a second, running twice as fast. Switching between normal and
fast is done by pressing the seconds button, which on a normal to fast switch,
syncs fast to normal. Holding the seconds button down clears the seconds, and
unconditionally switches to normal mode.

Incrementing hours and minutes works in both modes, although in the seconds
mode increments are not written to the RTC.